iot/shifu/examples/siemensPLCDeviceShifu/requirements.txt
requests
python-snap7
flask
